Extract storage into a separate class
I/O is too low-level to keep it with app logic or UI.
Maybe we'll want to support different storages in future
versions.

The only GUI part that's now storage-dependent
is the config dialog, but that's fine, as that dialog
is exactly what defines what kind of storage is used.