Add library stats and its serialization
To create global lessons, we'll need to keep track of when
a word was used last and how many attempts it took
to get it right. We'll use this information to select words
for a global lesson.

The first step is to introduce data structures and their
serialization. Then we'll need to actually compute,
save and load those.