Load, compute, update and save stats
For global lessons we'll need to know when a word
was learned last, how long it was before the previous
time, and how many mistakes were made.

To initialize this for the first time, let's assume that
the previous attempt was a day ago, just to have something
for the algorithm to work with. It should then increase/decrease
this interval accordingly, but it has to be implemented yet.