Refactoring viewFlashcards, startLesson, openFrame
viewFlashcards() doesn't work correctly now because
we need all flashcards for it, not just the top 30,
as for lessons in the current implementation.

To make it possible to change it, we refactor these
methods in such a way that the caller is able
to control how the data is prepared.

Next step: actually prepare the right data.