Ignore hidden dirs and files
Instead of ignoring just the .flakardia dir, it makes much more
sense to ignore all hidden files. For example, if a flashcard
set file is being edited with Vim, there'll be a swap file
which will break the parser.

Just ignore everything that starts with a dot. Of course,
it's more of a Unix conversion, but Vim follows it on Windows
as well. If we ever need to ignore more than just Vim swap files,
we'll just make a more sophisticated mechanism, for example,
based on name patterns or something like that.