Introduce randomness
When a group of words is learned, the words that fall into the same
number of mistakes category will then likely be repeated together
in one of the next lessons. This creates a sort of mental bucket
effect that makes these words deceptively easier to remember,
once you remember one word, the others pop up in the mind
automatically.

To fix this, introduce some sort of randomness that affects
the coefficient applied to the learn interval. By moving words
back and forth a bit in a random fashion, we eventually cause
them to be grouped in different lessons.