Refactor threading
Drop the service abstraction, as it was exposing threading
internals anyway (the indicator, the BGT/EDT parts...).

Make the EDT the main thread for most tasks to remove
those numerous ui {} blocks. Now most stuff runs in the EDT
by default.

Implement launchUiTask with a channel and a loop.
This lets us start UI coroutines from pretty much anywhere.
Running the loop in runBlocking ensures that the main thread
is kept alive and the app won't suddenly quit because no thread
is running before we show the main frame.

Hide that ugly indicator inside backgroundWithProgress.
Then we can refactor it.