Increase frame extra width
Sometimes it still wasn't enough, causing the horizontal scroll bar
to appear. We could fix the calculation, of course,
but given the Swing nature it's not an easy thing to do
and it would be a fragile fix working only until the next redesign.

Having an extra 50 pixels, on the other hand, looks nice and works reliably.