Fix frame width
When scrolling to selection, the list width is updated,
therefore the code in updateWidth() doesn't work
because it gets the new width. But the scroll pane
width stays the same, so essentially the code just
does "make the list as wide as it already is"
without accounting for the width of the actually
visible area.

Fix by using the scroll pane width
instead of the list width. And while we're at it,
include the extra width into the preferred sizes,
so it's always accounted for.