Save window location on move
Window closed/closing events aren't reliable.
For instance, on macOS windowClosing is only invoked
on mouse close, but not on our shortcut implementation.
In the latter case windowClosed is invoked,
but when pressing Cmd+Q neither is invoked.

Fix by saving location on move. To avoid doing it too often,
introduce a flow and debounce it with a 50 ms timeout.
The location will be saved 50 ms after the window has stopped moving.
This is reasonably fast.