Do not reinitialize the library if it hasn't changed
When the user opens the settings dialog, for example,
to change the font, it's very distracting that the library
is reloaded and reset to the root.

Change the implementation of the configuring function
so that it returns the new library and check in the manager
if the library is the same. Check it also in the presenter
and do not reload the entries unless the library has changed.

The library is a data class and so is its storage,
so comparing libraries is essentially comparing storage paths.