Update width after packing
Applying the new settings cause the frame to repack,
but packing causes the entry list to shrink to its preferred
size, without considering the extra width we apply in
updateWidth(). Fix by calling updateWidth() after packing.

An alternative approach would be to simply override
getPreferredSize() for the components with the extra width.
Then it would be accounted for during packing.
But we'd still need updateWidth(), as we call it when
the entries are updated, when we enter a dir.
We could call adjustSize() there, of course,
but then it could have other side effects besides updating
the width. For example, it could affect the height,
and the user likely doesn't want that (e.g. if they set the height
to their own preference). So let's keep the width separate for now.