Move all application logic to background
Now that we finally have a decent coroutine API,
this is a no-brainer, and it could be useful,
as it gives us a lot of freedom in the app implementation.

The app logic in question is the logic of the lesson,
library and the manager. Pure immutable data classes
are OK to be accessed from the EDT for now,
as it would be a useless overkill to enforce BGT access for them.