Use a mutex for model access
Now that we have moved all model access to the background,
it's automatically concurrent. That's not a good thing,
actually, as our models are mutable. This doesn't seem to be
an issue simply because all our operations are very fast
for the time being so they don't have much chance to actually
conflict with each other, but it's still a bad design.

Since all model access is required to be on background threads
now, the easiest way to enforce safe access is to introduce a lock
and ensure it's held prior to entering any background operation.
We could make it automatic, but sometimes it requires
granularity control. For example, when the lesson changes its
state, and then we save the stats, it's really a good idea
to ensure that the stats can't possibly change in between.
So we have two background ops (lesson access and stat saving),
but need a single lock session.

Introduce underModelLock() and use it everywhere for model access.
There should be no deadlock risk, as we only have a single lock
now, and we only use it for model access.