Fix an initial selection bug
The code assumed that the selected entry must be either
the up entry ("..") or a dir entry. But it could also be a file,
and in that case there would be no selection, because the code
didn't account for that.

Fix by performing a search for an entry with the specified path.
If there's no such entry, select the first one.

The new logic is a bit different from what the previous one
intended to do. In particular, if the entry is not found,
the first one will be selected. The previous implementation
assumed that nothing should be selected in such case.
It's not a very important difference, and it's not a bad thing
to ensure that there's always a selection.