Editing: the next step, implement card changes
The design is definitely going to be improved.
At the moment the mutable state relies on update
from the EDT. We likely need to move to request-based
state updates and encapsulate the state flow
into the presenter superclass.

Also, the relationship between the card presenter
and its state is a bit unusual: the state holds a reference
to the presenter, not vice versa. This is to enable
updates from the view, but it likely needs to be reworked.

No card additions nor deletions yet.