Migrate to sequential state updates
Async state updates started to cause troubles with the editor.
It's not easy to reason about the current state when it's mutable.
So order updates now using a channel.
Now all presenter updates come through that channel
and the previous state is always used as the base state.
In fact, it's no longer directly available to presenters.

The only tricky part was that selection handling in the manager.
Now that updates are delayed, it caused an infinite update
loop when the newly initialized list would have no selection,
and would immediately request the presenter to clear the selection.
Then the actual selection would arrive from the presenter,
the list would report it back to the presenter,
but then the null selection requested previously would finally arrive
and so on. To fix, do not send the selection back to the presenter
if it comes from the presenter to begin with. If it's the latest state,
then it's already synced, nothing to do. If not, then the latest
state will arrive soon anyway. Only selection changes
caused by user actions should be reported to the presenter.

Now the editor presenter looks much better. Replace that ugly
state--presenter reference with integer unique IDs that
are now used to identify individual cards even as they're edited.

Drop those epochs while we're at it. They were introduced
to avoid applying outdated updates, but at this point they're
not needed yet. We can re-add them if needed.