Clean up temp files
Ensure that there's always at least an attempt to delete
the temporary file. If it was created successfully,
then the chance of successful removing it is also reasonably high.
But the chance of successfully renaming it into the target
file is noticeably lower, as the target file could be read-only
or could be currently opened in some app.