Improve Enter handling
Enter at the beginning of the first line of a card inserts a new card.
Anywhere else in the first line it now moves the focus to the next
field, so after completing the question part, the user can instantly
move to the answer part.

From the answer part, now hitting Enter anywhere in the field
inserts the next card. It's more convenient this way,
as editing may finish somewhere in the middle due to some
corrections.