Fix focus and scrolling
When focusing a component, we want to make it visible.
The usual scrollRectToVisible() doesn't work, as it
has a special implementation for text components,
which is supposed to scroll the text within the component.
We don't need that, we need to scroll the entire editor
to make the component itself visible.
So we manually look for the viewport parent and ask it to scroll.

The whole thing is wrapped into invokeLater() to make sure
whatever validation requests we have issued have completed.