Add distinctUntilChanged() to state flows
Now that we're no longer using StateFlow to prevent
incremental values from being lost,
we need to manually filter repeated states.
Failure to do so would cause the initial editor state,
for example, to be applied again and again to the presenter
when the repeatitive saving bug was there.

In the absence of bugs it's not that important anymore,
but it won't hurt anyway.