Fix focusing and scrolling
Two issues: the component may not be showing when we want
to focus it, and it may not have valid bounds, which makes
scrolling impossible.

Fix the first issue by attaching a self-removing hierarchy listener
that trigger focus-and-scroll when the component is shown.

Fix the second issue by revalidating (just in case) and invokeLater
to ensure that the code is invoked again after revalidation.