Save the last state when the editor closes
Do it in the same loop the state is updated is to ensure
that it's indeed the last state and that save operations
are not concurrent.

Use a non-cancellable section to ensure that the save
operation is not interrupted because the view is closed
and the entire thing is cancelled.