Add some spellchecking
Support custom dictionaries because the built-in ones cover
just a few languages. To avoid the mess of dictionary
conversion, do it the easiest way possible: custom
dictionaries are just lists of words, and they go to the spell/
directory. That's it.

Spell checkers are registered when the components are showing
and unregistered when they're not.