Improve focus and caret handling on move / delete
When moving the caret using up/down, keep at home/end
if possible.

When deleting a card, adjust the focus and caret position
in the most natural way. Specifically, if a card is removed
using backspace, and it's not the first card, then the caret
is moved to the end of the previous card, so hitting backspace
again and again keep removing cards backwards.
The same for the delete key: if it's at the beginning of a card,
then hitting it again and again eventually removes the card
and continues onto the next one, from the beginning.