Refactoring: rename front and back
The two sides of the card were named the front and the back
in some places and the question and the answer in some others.
Turned out that the latter is easier to remember and to reason about.
Get rid of this ambiguity and use the question and the answer
everywhere instead.