Refactor and fix saveUpdatedFlashcard
Delegate to saveFlashcardSetFile.
This also fixes two issues.

First, warnings and errors from save results
were not properly handled because there was
some is-check that ignored the existence of
the SaveWarnings class and treated it as SaveError.
This is fixed by the delegation.

Second, throwing an exception when the card
is not found was a bad idea. The file could've been
changed externally and the card we're trying to rename
may no longer exist. In this case it's best to do nothing
than to throw an exception.