Refactor Lesson: use unique IDs
Now that we have tests for those flashcard updates,
we can refactor the Lesson class to reduce the dependency
on exact cards and instead use abstract IDs.

This refactoring allows to simplify greatly reasoning
about consistency of the state after flashcard edits,
as now IDs don't change so we only must update
the mapping and the correctness of the last answer.