Do not move to the next editor on Delete / Backspace
Some practical testing shows that it's not a desirable behavior.
It was added to enable deleting a lot of cards by holding
Delete or Backspace, but in practice cards are seldom deleted.
But they're often edited, and sometimes it's desirable to remove
everything in a single line, but not anything around it.
Holding Delete at the beginning or Backspace at the end
seems the most natural way of doing it, but with the current
behavior it starts removing characters from adjacent lines,
and that's not a good thing.